The dist target doesn't need hash or clean as prerequisites.
The "git archive" command pulls files directly from the repository in
.git/objects rather than the visible files.  Therefore neither hash nor
clean are required.  The notion of having the hash prerequisite came
from an earlier idea about making a file with hashes and copying that
into the tarball.  Thinking it over, I decided it was better to just put the
Git stuff into the Makefile by way of sed(1).