Allow compilation of sfrotz with clang
Add the -fPIC -fpic that seem to be needed when clang is used
as the compiler, and don't seem to hurt for gcc.  Also add commented
out CC=clang line to suggest what the user needs to do to use clang.
From one bug report it seems that OS/X uses clang even when gcc
is specified as the compiler.