Fix two strcpy() calls in curses and needed to redefine os_read_file_name().
The os_read_file_name() function has been a source of odd looks for a
long time.  The biggest of the reasons to look at it funny was how it
passed a string back to the caller by reference without clearly stating
how big of a buffer was being written to.  It was always FILENAME_MAX,
but still, it looked like a specialized strcpy().