Revert "Fixed problem with garbage .aux filenames."
According to Andrew Plotkin, Vlaviano had it wrong in
https://intfiction.org/t/unix-frotz-issue-with-extended-save-and-restore/9816.
The correct grammar for calling @save and @restore is "The third
argument should be the address of a byte array containing the filename.
(Byte zero is the name length, followed by the characters in ASCII.)".
Vlaviano mistakenly thought that the third argument is supposed to be a
literal string.

This reverts commit 32d85186baeaf817873049f7937e28ffd37a4e39.