Xfrotz compiles and runs zcode.  Problems with the screen and save-restore.
On starting up after, the game screen will usually be mostly black except
for a small white block where the prompt is.  This hides the exposition and
first lines of the game before the first prompt is presented.  At this
point, pressing ^C will sometimes fix the problem.  Subsequent runnings of
xfrotz might then look okay.  I suspect the screen problems have to do with
some variable(s) somewhere not being initialized.  Years ago, this was
somehow not a problem.

There is also trouble with saving and restoring: xfrotz segfaults
immediately.