DOS port now compiles and works.  Issues #8 and #58 need revisiting.
Loading an executable in a Blorb file works, but very slowly. Audio and
graphics support through Blorb are not yet attempted, though sometimes
old-style works.  For instance, in Sherlock, blowing the sound of a cab
approaching is heard (blow the whistle on the first move).  However,
playing the violin after you go upstairs is not.