Fix problem of os_fatal() trying to write to SDL window when SDL not active.
If we have a showstopping problem right from the start, like trying to
load a file that's not a Zcode file or Blorb file that lacks a Zcode
chunk, then Frotz will call os_fatal() to complain about it and then
exit.  But sfrotz's version of os_fatal doesn't check if SDL is active
or not before calling functions that depend on SDL.  The first of these
is os_set_text_style().  This leads to a segfault when a call of
current.font->getglyph(current.font,c,1); dereferences getglyph, which
does not exist because current is not properly initialized yet.