Fixing assert due to failure of src_process
Based on discussion with @borg323 src_process was leaving the
input data pointer at an old value if it ran out of input data and
completed the output buffer in the same call.  Fixed it by setting
the input data pointer back to the beginning, but since the number
of input frames was set to 0, it wouldn't read any data from it.