Handle timed input and sound sequences at the same time
The earlier change to make sound sequences work correctly by periodically
calling os_tick by setting a maximum wait when the timeout was set to 0
(wait forever).  If the timeout was set to a non zero value, then os_tick
would not get called until the timeout completed.  This change determines
if the timeout is larger than the max wait time, and if so, breaks thw
timeout into a series of maximum wait time waits, with calls to os_tick.