Allow for alternative Blorb file at command line for curses, sdl, and dumb.
Doing this for the SDL interface is proved a bit troublesome because
it's not very clear where stuff is being set and where not.  After the
audio overhaul release, I want to overhaul the SDL interface code to
make it as closely resemble as feasible that of the curses interface.