Avoid race conditions when doing a parallelized buid.
If the -jX flag is used to indicate that X CPUs should be used to build
Frotz, there is a slight chance that the definitions files:
src/common/defs.h, src/curses/ux_defines.h, and src/common/git_hash.h;
will not be complete before the build process starts on the rest of the
source files.  If this happens, the build fails.