First stab at V6 semi-support through Blorb in curses.  Things are distorted.
Right now, the code detects PNG and JPG images.  However, stuff seems to
be done only for Infocom's games and the blorbs made up for them.  Even
then, the results are distorted.  This may be because rectangle chunks
aren't being processed and the code depends on some chunks that aren't
found in post-Infocom V6 blorbs.  Still, here's something to play with.