Remove R from getopt that isn't being used for anything yet.
I just noticed that the -R flag for restricting where sfrotz is allowed
to read and write is not implemented.  This feature isn't quite as useful
for the SDL interface as it is for the curses and dumb interfaces, so
I'll give this a 2.46 milestone.  If anyone reading this issue feels like
implementing this now, that's great and I'll merge it when you're done.