Remove version.c and put build date back into git_hash.h.
I was never quite satisfied with the way I did version.c.  I seem to
recall someone telling me it was improper to put the build date into
git_hash.h.  After a couple months of this, I looked again and realized
that I picked an irritating solution -- Frotz will always be rebuilt
when doing "make install".  The difference between when version.c is
created and when someone is done monkeying with the build process does
not seem to be of much consequence.