Don't add $(PREFIX)/include and $(PREFIX)/lib to toolchain paths
PREFIX is only applicable at install and runtime. The headers and
libraries we're building against may not be in the same location,
particularly if we're cross-compiling. Adding -I/usr/include when
cross-compiling will likely break the build.

It is generally not necessary to add these paths anyway as toolchains
are already configured to look in the appropriate places for system
headers and libraries. There may be exceptions but these can be dealt
with via CFLAGS and LDFLAGS. Perhaps this is necessary for MacPorts
but I'm not familiar with it.