First step of changing "screen height|width" to "text height|width".
The purpose of this is to distinguish between text dimensions and screen
dimensions.  This stems from a problem with specifying the dimensions
for sfrotz which results in graphics in a V6 game not being properly
scaled to the window dimensions.  For sfrotz, we will have -w and -h
specifying the dimensions of the text as is with the dumb and curses
interfaces.  To this we will add -W and -H flags to specify the window
dimensions.  That pair of parameters will alter whatever is being
altered when the sfrotz is manually resized.