Delete references to special circumstances for setting text width.
Particularly with "again", which referenced something long since
deleted.  For the SDL interface, manually setting the text width could
be a very useful setting.  For the dumb and curses interfaces, one
generally wants the width to the be the same as the terminal.  Frotz has
always been able to automatically deal with the width of the terminal.