Rewrite Blorb file loading to allow for more robust file handling.
Previously, the way Blorb files were opened was a bit haphazard.  This
led to segfault problems when trying to load a naked Zcode file that has
an acompanying Blorb file.  Typically the Blorb file is found in the
same directory as the Zcode file.  The segfault would occur because the
path for the Zcode file was not prepended to the generated Blorb
filename.  When I wrote Blorb support for the curses interface, I didn't
have/cause this problem.