Fix issue with save files on DOS
DOS in large mode can only address 16 bit chunks of memory.  When
writing the stack in the save file, the save and restore routines
look at the zcode to sanity check the number of arguments on the
stack.  The code size can be more than 64K, and in that case, the
wrong value can be stored in the save file.  Similarly when loading
a save file stored in linux, the check will fail because the value
read from the save file does not match the value at the offset when
it is used modulo 65536.