Fix processing of save file command line argument
On init in restore mode, frotz copies `tmp_save_name` to
`save_name`. The number of bytes copied does not match the amount of
memory allocated.

The size of copied memory block is based on story name and not save name.

If the save file name is longer than story name, then only part of the
file name will be copied, and the restoration almost certainly
fails.

If the save file name is shorter than story name, then too many bytes
are copied, potentially corrupting memory. Restoration is also likely
to fail unless a zero byte happens to be present at the right place of
the incorrectly accessed memory.