Add a -f <format> mode for dumb interface -- just IRC and normal modes for now.
This makes the dumb interface more interesting when used as the engine
of an IRC IF-bot.  To start with this, I manually applied the patch from
Philip Lafleur, which is found at
http://ifarchive.org/if-archive/infocom/interpreters/frotz/frotz-2.43-irc.diff.gz

There were some problems in getting this patch to coexist with the
normal mode, so I rewrote src/dumb/doutput.c to switch from representing
an individual cell (character + foreground color + background color +
style) as a packed integer to a structure.  I feel that this rewrite
makes the source file easier to understand and will avoid potential
portability problems later on.  I'm thinking of the upcoming PDP10 and
PDP11 ports in particular.