Add snavig.pl for converting source files for compilation on TOPS20.
This is a conversion of Adam Thornton's gnusto-frotz-tops20[1] script which
clones a Frotz repository and converts files to a form suitable for
compiling on a TOPS20 machine.  snavig.pl is essentially the same as
Adam's script except that the latter is simplified to work files already
present.

The name of this script is taken from a spell of the same name from
Infocom's game "Spellbreaker".  It's purpose is to "Change an object's shape".
See http://www.ifwiki.org/index.php/Spells

[1] https://github.com/athornton/gnusto-frotz-tops20