Remove "Build date:" from -v option output.
In the interest of making sure that the executables a compiler produces
come out exactly the same every time they're built, I'm removing the
"Build date:" line from the -v option output.  Knowing when the
executable was built can be handy, but is ultimately not as useful as
knowing when the latest commit was made and what its hash is.  Those do
not change each time the executable is built.

See https://reproducible-builds.org/