Rework Makefile to better handle curses library requirements.
Sometimes ncurses is built with termlib split off into a different
library.  Sometimes it's integrated with ncurses.  For the former case,
we need to use -ltinfo in additon to -lncurses or -lncursesw.  For the
latter, all one needs is -lncurses or -lncursesw.  Exactly what's needed
per installation isn't exactly clear, so these changes should satisfy
whatever machine for which this matters.