Allow list of directories in path
Allow multiple search directories in the path environment variables
and config files.  It looks like this capability was once their, but
was lost over time.  The sdl port allowed a list of paths in os_path_open,
but os_path_open was not being used.  Also the sdl implementation had a
bug where the path might have garbage at the end.