Preliminary fix for problem passing praxix tests when UTF-8 not used.
The problem is that the oe and OE ligatures (ZSCII 220 and 221) aren't
actually in the ISO/IEC 8859-1 (aka Latin1) character set.  Prior to
adding UTF-8 support, the zscii_to_latin1[] table mapped to 8-bit
characters with 0x00 appearing where these ligatures should have been.
For Latin1, those positions are occupied by 1/4 and 1/2 characters.  Now
that UTF-8 is supported, this table is made up of 12 bit characters.
Because 0x153 is nine bits long, the ninth bit is stripped off, which
yields 'S' and 'R' respectively.  This fix puts Latin1 codes for these
fraction glyphs where they belong.  Now Frotz will pass the praxix
streamtrip and streamop tests, but will return the fraction glyphs
instead.  Frotz didn't pass these tests when 0x00 was in there.

The ISO/IEC 8859-15 (aka Latin9) character set replaces the 1/4 and 1/2
glyphs with the OE and oe ligatures.  So maybe we should have additional
tables?  Or perhaps map the ligatures to 0x00 and 0x01 and then have
unicode_to_ascii() to check for this and return the correct code.