This doesn't work, but maybe someone can see something I don't.
When the "-f ansi" option is used under TOPS20, the first instance of
the status line is printed with a spurious newline after the 69th
character is printed.  Neither the -w nor -S options seem to affect
where the spurious newline appears.  Clearly something is being done
wrong in the two blocks where foreground and background colors are set.