Revert "This doesn't work, but maybe someone can see something I don't."
This reverts commit a6cb4878a53df6e4e51518899a8058aa7a24126c.

These changes are not necessary, at least not yet.  Maybe they can be
revisited for cutting memory use.  The whole problem of spurious
newlines I now understand is from TOPS20 itself.  It inserts a newline
automatically after every Nth character where N is the width of the
terminal as set with the "TERM WIDTH" command.  To turn off this
behavior, one should type "TERM WIDTH 0" before running dfrotz.  This
can also be accomplished through the jsys() system call.  More
practically, "TERM WIDTH 0" should be put into the system or users'
LOGIN.CMD files so that it is run automatically.

I found a program that creates a real-time analog clock on assorted
systems terminals, including TOPS20.  That's where I discovered the
jsys() call.  The program is "crock.c" and can be found at
https://www.hack.org/mc/software/crock.c.  The following block appears
to do the job.

\#ifdef KCC
     ac[1] = _PRIOU;  /* 1/ .PRIOUT   */
     ac[2] = _MORLW;  /* 2/ .MORLW   */
     jsys(MTOPR,ac);  /* MTOPR%    */
     te_width = ac[3];  /* returns 3/ width  */
    /* set infinite terminal width: */
     ac[1] = _PRIOU;  /* 1/ .PRIOU   */
     ac[2] = _MOSLW;  /* 2/ .MOSLW   */
     ac[3] = 0;   /* 3/ 0    */
     jsys(MTOPR,ac);  /* MTOPR%    */
\#ifdef OPTIMIZER
     ac[1] = _PRIOU;  /* 1/ .PRIOUT   */
     ac[2] = _MORLL;  /* 2/ .MORLL   */
     jsys(MTOPR,ac);  /* MTOPR%    */
     te_height = ac[3];  /* return 3/ height  */
\#endif
\#endif