use ncursesx.x-config instead of pkg-config
tl;dr, this commit searches the path for a ncursesx.x-config file and
runs the first one it finds to get the flags for curses.

ncurses is available by default. There's no ncurses.pc file for
pkg-config, however. So, even if pkg-config is installed, it can't be
used to get the flags for ncurses (unless PKG_CONFIG_PATH is set, which
complicates things further).

Homebrew has a ncurses installed, which *does* have pkg-config info, but
it is "hidden" so it doesn't interfere with the system ncurses (see
`brew info ncurses`). I don't use macports or fink, so I'm not sure how
they handle a separate installation of ncurses.

In short, if ncurses is installed somewhere, and its config program is
in the user's $PATH, this will find it and use it. I think.