@set_true_colour handling looks like how ncurses does it.  dfrotz works too.
The next step from here is to get the "rainbow" in the spec11 praxix
test to really look like a rainbow.  As-is, I get one block of red, 26
blocks of red, 26 blocks of cyan, one block of blue, and 26 block of
purple.  This fills 80 columns.

Then there is the matter of the "Recommended:" and "Interpreter:" lines.
There should be black, red, green, yellow, blue, purple, cyan, white,
light grey, medium grey, and dark grey.  The "Recommended:" line shows
up with cyan where blue should be.  All three greys are display as
white.  For the "Interpreter:" line, all three greys display as black.

Sfrotz already correctly did all of this before being incorporated into
this repo.