Fix occasional rendering bug.
As LadyHavoc explained in an [SDL forum post](https://forums.libsdl.org/viewtopic.php?p=42023#42023),
"You may also have to do more than one Present before any new image is
shown to the user, due to the driver telling the GPU to present an image
multiple frames after it was issued by the application, usually as part
of "double buffering" or "triple buffering", or as a result of the
render-to-texture behind the scenes in a compositing desktop environment
(which many are today, even on Android and iOS). So in general you want
to spin your main loop and constantly refresh..."

The problem was seen with Infocom's *Arthur: The Quest for Excalibur* on
macOS. Several images are shown at the beginning of the game, and the
images would not be shown some small percentage of the time (maybe up to
5 percent of the time, depending on computer activity and other random
factors).