Using some simple markup to mark the beginning and end of stuff.
Need to figure out how to get rid of the "Ok." chatter from saving and
restoring.  That's generated by the game, not the interpreter, so this
may get funky.  I might try something like forcing printf() to go
through a wrapper that causes nothing to be printed if the game tries to
print something we don't want.  Basically, I need to filter out the
response from the first command (RESTORE) and the last one (SAVE).

Each invocation should end with a prompt.