Curses can't change the cursor color in an instant.  Trying to compensate.
In order to get Frotz to do the Right Thing when a color change is made,
it needs to emit a raw ANSI terminal sequence to do so.  I found
discussion on the subject at
https://stackoverflow.com/questions/18425103/how-do-i-change-the-cursor-color-in-ncurses-forms
A brief explanation of the sequence in question in at
https://invisible-island.net/xterm/ctlseqs/ctlseqs.html.

Basically, it's this:

\033}12;#rrggbb\007

The characters rr bb and gg are replaced with hex representing red,
green, and blue.  Right now I'm just checking if $TERM is xterm to
determine if it's okay to emit this sequence.  I want to expand this to
a list of acceptable $TERM values known to do the right thing.

Because the curser appears as a bold (or brighter) of the specified
color, I need to multiply the detected color values from color_content()
with some factor to get what should be there.  Unfortunately dividing
the cursor color as determined by a color picker by the value I get from
color_content() gives me a value (1.4167) that when multiplied as a
float with a short, I get something too small and thus too dim a cursor.
By trial and error, I found that the proper value is 1.48.  I strongly
suspect this value will cause wrong colors if someone has different
color definitions.

If I go through with this, it may require a command line option to turn
it off in case it annoys people.