Added -E option to control how EMPHASIS_STYLE is handled.
Infocom designed the emphasis style to be rendered as italic (default)
or underline depending on the capabilities of the target computer.  A
third option of doing neither is also available.

Italic support was rare and spotty in the early days.  Underline was
much more common.  It wasn't until version 5.10 (change added September
13, 2013) that ncurses got support for italic.  Some platforms didn't
support either, notably Apple II and DOS.  In those cases, the
interpreter or game would render emphasised text in all-caps.  For this
addition, an all-caps is not done because to add italic was a simple
matter of passing a flag to ncurses.  To do all caps will require a bit
more work.

A proof of concept for italic support was presented by Torbjörn
Andersson, which I then refined into what is seen here.