Changed license notices for X11 interface to GPL2.
When the X11 interface code was first written, the rest of Frotz was
still under its old "no commercial" license.  It read as follows:

Frotz is freeware. It may be used and distributed freely provided
no commercial profit is involved. (c) 1995-1997 Stefan Jokisch

Later, when Stefan handed me custody of Frotz in 2000, he changed the
license to GPL2 or any later version.  I asked Daniel Shepler, author of
the X11 interface code, in January 2020 if he'd be willing to change the
license on his code to GPL2 or later as well.  He agreed to this, but
hadn't the time to make a commit himself to this effect.  Therefore I am
making the commit myself.