Simplify: zgetopt does not really need to be cdecl on MS-DOS.
There is no real need for zgetopt to follow the cdecl
calling convention, even on MS-DOS, so I have removed the
requirement.

(The code for enabling/disabling cdecl for zgetopt was a bit
glitchy in any case: src/common/getopt.c did not include
"frotz.h", so MSDOS_16BIT was not correctly defined even for
DOS hosts.

I encountered this issue while porting the DOS code to the
Open Watcom compiler toolchain.)