Now can start, save, load, etc; but the save file crashes normal mode.
These changes suggested by Borg323 allow Dumb Frotz to start up in Bot
Mode with no save already there, play a move, save, restore, play
another move, and so on indefinitely.  However, the resulting Quetzal
file will somehow crash any flavor of Frotz when not in Bot Mode.  This
seems to call for an investigation into exactly what is being saved in
the Quetzal file for normal mode and Bot Mode and how the state of the
Z-machine differs from a normal save versus Bot Mode.  Once this is
understood, I hope it will be a simple matter to fix up things prior to
doing a save.

The above is valid for zork1.z3.  I tried with sherlock.z5 and
discovered that the game would not progress.  If I loaded the resulting
Quetzal file with the curses interface Frotz, the result was "[What?]",
which is what happens if you press Enter without typing anything else.
Then the banner at the top of the terminal is corrupted.