Worked around a problem wherein LGOP has a really long prompt.
When you SMELL something the first time in Leather Goddesses of Phobos
(LGOP), you're told to scratch and sniff spot number N and then hit
RETURN or ENTER to continue.  On an 80-column screen, that leaves 8
characters left to type something.  Possibly because a V3 game has only
8 characters of dictionary resolution, Infocom's interpreter that comes
with LGOP places the cursor right after this.  Frotz has always checked
to make sure there is at least 10 characters after a prompt and if not,
issue a newline before presenting the cursor.  Offhand I can't remember
which of Infocom's games besides LGOP have really long prompts like
that, so for now I think it's okay to reduce this check to 8 characters
for just LGOP or maybe any V3 or below.

Since the dumb interface sets a default width of 75 columns, this
minimum space needs to be reduced to 4.  There seems to be no good
reason to keep it like that, so I've increased the width to 80.

A sidenote, the Apple II version in 40-column mode also puts the cursor
at the end of the really long prompt.  When you type off the edge of the
screen, it continues on the next line rather than repeatedly change the
character at column 40.  You can backspace to erase the whole command,
but the cursor won't go back up to the previous row.