Corrected unreachable code problem in sfrotz that prevented -I from working.
Sfrotz first checks .sfrotzrc for settings, then checks the command line
flags for any overrides.  The way the interpreter number was handled was
that if f_setup.interpreter_number is 0, then default to 4 (for Amiga).
That would always be set to 4 because the only other way to set
f_setup.interpreter_number is at the command line... which is executed
AFTERWARDS.  So, now we'll just always set z_header.interpreter_number
when parsing .sfrotzrc and set it again in parse_option() if the -I is
used.