Added to sfrotz an -M option for resolution multiplier.
This has the effect of multiplying the resolution of the SDL window in
sfrotz by the integer supplied.  Size of the type is not affected.  The
values of z_header.screen_width and z_header.screen_height are also
multiplied.  This is not desirable for Infocom's four V6 games because
those values in the Z-machine header are how the games decide which
graphic interface is being used for IBM and Macintosh.  The objective
now is figure out how to give the Z-machine what the resolutions the
four V6 games expect and then adjust the final scaling to whatever we
want.