Fixed problem with Frotz not entirely following the spec for timed input.
borg323 noticed this after I reported segfault trouble when timed input
is used in a certain way.  The certain way was a demonstration of
trouble that turned out to have been fixed by the time of Frotz 2.50,
but this oddity remained.  We both believe this may be part of the
problem.

Below is a passage from the Z-Machine Specifications version 1.1 section
15 talking about the @read opcode:

  In Version 4 and later, if the operands time and routine are supplied
  (and non-zero) then the routine call routine() is made every time/10
  seconds during the keyboard-reading process. If this routine returns
  true, all input is erased (to zero) and the reading process is
  terminated at once. (The terminating character code is 0.) The routine
  is permitted to print to the screen even if it returns false to signal
  "carry on": the interpreter should notice and redraw the input line so
  far, before input continues. (Frotz notices by looking to see if the
  cursor position is at the left-hand margin after the interrupt routine
  has returned.)

The critical sentence is this:

  If this routine returns true, all input is erased (to zero) and the
  reading process is terminated at once.