Changed len's type from int to size_t.
In os_read_line()... It seems that due to saving the return value of
zcharstrlen() to an int rather than size_t (which is the type it
returns), somehow that variable (len) got rolled over to nonsense.
Making len a size_t seems to fix this.  There are other parts of
os_read_line() that look like they could use some casts to make sure
similar problems don't happen again, but I'll leave that for the next
commit.