Fixed problem with @aread being used inconsistently.
cwp->x_cursor should never be less than os_string_width(buf).  If this
does happen, it means that the width of the buffer is believed to be
negative: obviously wrong.  Doing this results in a segfault.  I haven't
heard of this happening in the wild, but a program given to me intended
to demonstrate another bug triggered this one.

This commit catches this bad behavior and causes Frotz to abort with a
fatal error message when it happens.

Here's a minial working example to demonstrate the problem:

! A test for inconsistent input buffer.
! Code copied from Etude.
! Uncomment first line for correct behavior.

Switches xv5;

Constant INBUFSIZE 80;
Array inbuf -> INBUFSIZE;

[main inbufvar ix;
!  print "Preload> Given";

  inbuf->0 = (INBUFSIZE-3);
  inbuf->1 = 5;

  inbuf->2 = 'G';
  inbuf->3 = 'i';
  inbuf->4 = 'v';
  inbuf->5 = 'e';
  inbuf->6 = 'n';

  inbufvar = inbuf;
  ix = 0;
  @aread inbufvar ix;
];