Fixed off-by-one error in @aread fix.  Error described below.
From Borg323:
A final thought: A more robust way to detect this case would be to keep
track of the last line of text printed and compare the buffer with the
last characters printed to see if they agree.  But this is probably
overkill.

The above comment was meant for future reference, this is a game bug and
the only bad case is currently (mostly) being caught. I did some
additional testing, and the only bad behavior I see with other
inconsistent cases is corrupting the text on screen up to the left
margin (depending on the input buffer size). However it seems I had an
off-by-one error, the test ought to be
cwp->x_cursor <= os_string_width(buf), I had wrongly assumed x_cursor
starts from zero.