Experiment with adding a delay for each instruction
This adds a short delay for each iteration of the interpreter.  This
is to test out the idea of "throttling" frotz to work around an issue
in Bureaucracy.

The change in the initialization of curses to turn on immedok() was
needed, otherwise the burffering would cause a several second delay,
and then all the lines would be printed (especially noticeable if you
"RUN PLANE.EXE". Sdl frotz does not suffer from this.  dfrotz seems to have
several issues with the display (need to hit return twice for each computer
input.  xfrotz has some problems as well.