Fixed problem with garbage appearing after scrolling through command history.
This problem manifested if you give a long command followed by a short
one.  After typing in those commands, if you use the arrow keys to
scroll up and down through the command history, garbage will appear
after the short command up to the length of the longer command.  This
seemed to boil down to scrnset() not being able to correctly handle a
space when USE_UTF8 is defined.  This is described thus:

In the row of the cursor, set n characters starting at start to c.

Since scrnset() is always called with a space, there's no need for it to
process any other character.  I therefore changed the function to
"spaceset()" to reflect how it's actually used and always have mvaddch()
write a space.  This might not be entirely correct, but UTF-8 handling
seems to be undamaged.