Fix compilation on systems without pthreads
The pthreads are only required if the SOUND_TYPE is set to "ao", so it
should be possible to compile Frotz also on systems that do not have
the pthreads library. However, compilation currently fails there since
ux_audio.c tries to include pthreads.h anyway. Fix the problem by
moving the inclusion of pthreads.h into the "#ifndef NO_SOUND" section.