Fix issue with expansion of abbreviations with -x
In some cases, the code that encodes the abbreviations can encode data
past the the end of the replacement.  The loop encodes 0s as padding,
but depending how the data are aligned in memory, since there is only
one zero added to the replacement, the loop can index past that, making
it so that the replaced word has garbage in it.

This was originally reported in the intfiction.org playing forum.