Get pictures for X11 from blorb file
This is a replacement for the original pull request.  It merges in
recent changes, and improves the speed of drawing.

Incorporated some ideas from http://cwikla.com/X/articles/txa/xcc.2.html
to improve the speed of drawing in X when using blorb pictures.  Basically
this means not allocating colors in truecolor mode, but using the predefined
color mask, and when not truecolor, to create a color map that is a standard
color cube, and mapping all colors to that cube.  For convenience a 6by6by6
color cube is used, so that it will fit in an 8 bit psuedocolor display.