Updated ChangeLog with DOS Frotz's Amiga font modes
Where Frotz was supposed to print the oe or OE ligatures, it was instead
printing 1/2 and 1/4 as single characters.  Several other glyphs from
Latin-1 don't appear in the DOS system fonts.  It is, however, possible
to display these glyphs if the Amiga display mode is specified with the
-d5 flag.  Unfortunately, the fonts supplied with Frotz still showed the
oe and OE ligatures as 1/2 and 1/4.  It may be possible to correct this
error with MCGA and EGA display modes.