Fixed problems with loading old-style graphics info and drawing pixels.
Thanks goes to @tkchia, who converted Turbo C idioms to ones that Watcom
C likes.  When I noticed that V6 graphics were not loading, @tkchia
noticed "a silly error on line 110".  Fixing that allowed graphics data
to be loaded.

The second error fixed was more subtle as it was buried in a chunk of
assembly that was mostly rewritten from the Turbo C equivalent.  There,
moving the "display" int into the SI register resulted in junk because
that int was in some other segment.

One way to fix this was to change this line:
      mov si, display
to this:
      mov si, ss:display

What was needed is the correct segment override to get the current
display mode enum into the SI register.  The second approach, which is
used in this commit, is to declare a new local variable and put the
current display mode enum there.  The compiler will then handle the
referencing automagically.