Added dparams.txt from the old bcparams.txt.
In case anyone hasn't guessed, these params files are for relieving
people of the need to move stuff round and round and round whenever a
new option is added.  Just add the new option in the order it needs to
go (alphabetical) and run src/misc/twocol.pl on dparams.txt and you get
something nicely formatted to paste into the Frotz source code.