Ensure that a sample plays to its end before halting playback.
Lurking Horror with sound was released only for the Amiga.  That version
assumes that the interpreter is very slow, so it fires off several sound
effects in one round assuming there will be time to play most of each
one.  What happens on a fast modern machine is the sound starts and is
immediately stopped, so they don't appear to play at all.

See the "Remarks" section in
http://inform-fiction.org/zmachine/standards/z1point1/sect09.html

When the SDL Frotz code was originally written or converted to SDL 2,
this caveat was forgotten.