Added --xscale and --yscale for EGA V6 graphics.  Split m_v6scale to x and y.
This entailed switching from the Frotz-contained zgetopt() to the
OS-provided getopt_long().  There were no sensible single-letter flags
available for this.  Given that I don't anticipate SDL Frotz being built
on any machine that doesn't have getopt_long() and/or getopt.h
available, this is a reasonable choice.  If someone wants to submit a
patch to implement long options, I'm open to adding it