Reenabling old-style graphics for X11 interface.
This is for two reasons.  One is just for grins.  The other is for
investigating how to properly process EGA and CGA old-style graphics
files.

To get EGA properly framed, it seems that the thing to do is find
variables like gtable[i].width, gtable[i].height, data->width,
data->height, and divide by two.  Fiddling around with the eight places
I found in x_oldpic.c to do this, I managed to get width okay, but not
height.  I'm having a hard time identifying what to alter and what to
leave alone or if I need to look at other variables too.