Fixed trouble with z_header.user_name clobbering Inform version when V6.
Infocom stored a username in the same place that Inform stores its
version number.  Somehow it was discovered that the Z-machine would set
the username from the TOPS20 or Unix development environment, but it was
unknown at that time that any game made any use of it.  With the public
release of the Infocom source code, it was discovered that "Hollywood
Hijinx" used this functionality to enable a debug mode.  Because of this
clobbering when running V6 Z-code and that no other game seems to use
this, it's safe to completely ignore the username stuff.

Maybe it can be brought back in a roundabout way to put "Hollywood
Hijinx" into debug mode using the hotkeys.