Clarified a bit on the os_fatal() and os_warn() functions that use va_list.
The dumb interface does not use va_list printf functions because I'm not
entirely sure if the most primitive of C compilers will support them.
For this reason, the core must not call os_fatal() or os_warn() with
formatting.  For all other interfaces, formatting may be used as long as
the implementations of os_fatal() and os_warn() are written to support it.

It should be noted here that currently only the SDL interface uses os_warn().