Preparing for Z-machine standard 1.1 of not prompting on save/restore.
Copied from the Z-machine Standard 1.1 document:

As of Standard 1.1 an additional optional parameter, prompt, is allowed
on Version 5 extended save/restore.  This allows a game author to tell
the interpreter whether it should ask for confirmation of the provided
file name (prompt is 1), or just silently save/restore using the
provided filename (prompt is 0).  If the parameter is not provided,
whether to prompt or not is a matter for the interpreter - this might be
globally user-configurable.  Infocom's interpreters do prompt for
filenames, many modern ones do not.