Now shortens includes if necessary.
TODO: Guard against collisions caused by filename shortening.  For
example, if we shorten "foobarbaz.h" to "foobar.h", this will collide
with the existing "foobar.h".  The shorten_filenames() sub has a
mechanism for doing this.  I need to translate its sed-specific behavior
to the new way of doing things.