Ensure that symbols after #define is counted.  Doesn't work all the way.
Somehow, the symbols after an #ifndef are detected, but not replaced.
Even though the symbol to be replaced is in the symbol table. Looking
further, I see that the
s/\K\b($oldsymbols)\b(?=.)/$transformations{$1}/g; substitution is the
problem with the replacement failure.  I need it to capture more after
the match, but last time I tried to extend the (?=.) stuff, Perl didn't
like that.  More investigation and experiments are in order.